Extract TypeMappingHelper from TreeMapperPHPStanExtension
